+++ src/ncas_amof_netcdf_template/tsv2dict.py
@@ -319,7 +319,6 @@ def instrument_dict(
 def product_dict(
     desired_product: str,
     platform: str = "",
-    instrument_loc: str = "",
     deployment_loc: str = "land",
     use_local_files: Optional[str] = None,
     tag: str = "latest",
@@ -331,8 +330,6 @@ def product_dict(
     Args:
         desired_product (str): name of data product
         platform (str): location or observatory of instrument
-        instrument_loc (str): [DEPRECATED - use platform instead] location or
-                              observatory of instrument
         deployment_loc (str): deployment mode, one of 'land', 'sea', 'air',
                               or 'trajectory'. Default 'land'.
         use_local_files (str or None): path to local directory where tsv files are
@@ -344,23 +341,6 @@ def product_dict(
         dictionary of all attributes, dimensions and variables
         associated with the named data product.
     """
-    if platform != "" and instrument_loc != "":
-        warnings.warn(
-            "Both platform and instrument_loc are used, using platform."
-            " instrument_loc will be removed from version 2.6.0",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-    if instrument_loc != "":
-        warnings.warn(
-            "instrument_loc is deprecated, use platform instead"
-            " This option will be removed from version 2.6.0",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        platform = instrument_loc
-
     common_dimensions_url = values.get_common_dimensions_url(
         use_local_files=use_local_files, tag=tag, loc=deployment_loc
     )